:truck: Move file to ADT
